Massive HTTP performance improvements
We have enough memory for 1024 bytes to be a reasonable download size
which results in 8x less operations to download a file. I can't figure
out what the 50ms delay was for, but I suspect it was needed back when
the I2C between the CC3100 and STM32L4 was flaky due to running at the
wrong speed - I can't find any problems with it now.

These changes make the randomcat app load an image in 5s instead of 30.